DEV : creation of likelihoods and typing modules
