+++ src/relife/lifetime_models/__init__.py
@@ -51,28 +51,39 @@ Likelihoods
 
 """
 
-from . import (
-    _base,
-    _distributions,
-    _non_parametric_models,
-    _parametric_regressions,
-    _semi_parametric_regressions,
+from ._base import ParametricLifetimeModel
+from ._distributions import (
+    EquilibriumDistribution,
+    Exponential,
+    Gamma,
+    Gompertz,
+    LifetimeDistribution,
+    LogLogistic,
+    Weibull,
 )
-
-_non_parametric_api = [
-    "KaplanMeier",
-    "ECDF",
-    "NelsonAalen",
-]
+from ._non_parametric_models import ECDF, KaplanMeier, NelsonAalen
+from ._parametric_regressions import (
+    ParametricAcceleratedFailureTime,
+    ParametricLifetimeRegression,
+    ParametricProportionalHazard,
+)
+from ._semi_parametric_regressions import SemiParametricProportionalHazard
 
 __all__: list[str] = []
-# __all__ = [
-#     "CoxPartialLifetimeLikelihood",
-#     "EfronPartialLifetimeLikelihood",
-#     "BreslowPartialLifetimeLikelihood",
-# ]
-__all__ += _base.__all__
-__all__ += _distributions.__all__
-__all__ += _parametric_regressions.__all__
-__all__ += _non_parametric_models.__all__
-__all__ += _semi_parametric_regressions.__all__
+__all__ += ["ParametricLifetimeModel"]
+__all__ += [
+    "EquilibriumDistribution",
+    "Exponential",
+    "Gamma",
+    "Gompertz",
+    "LifetimeDistribution",
+    "LogLogistic",
+    "Weibull",
+]
+__all__ += [
+    "ParametricAcceleratedFailureTime",
+    "ParametricLifetimeRegression",
+    "ParametricProportionalHazard",
+]
+__all__ += ["ECDF", "KaplanMeier", "NelsonAalen"]
+__all__ += ["SemiParametricProportionalHazard"]
