+++ src/relife/lifetime_models/_non_parametric_models.py
@@ -8,12 +8,6 @@ from optype.numpy import Array1D
 
 from relife.lifetime_models._base import plot_probability_function
 
-__all__ = [
-    "ECDF",
-    "KaplanMeier",
-    "NelsonAalen",
-]
-
 
 class NonParametricEstimation(NamedTuple):
     timeline: Array1D[np.float64]
