+++ src/relife/likelihoods/__init__.py
@@ -1,0 +1,3 @@
+from ._lifetime_likelihood import LifetimeLikelihood
+
+__all__: list[str] = ["LifetimeLikelihood"]
